+++ src/main/java/kitchenpos/domain/Menu.java
@@ -4,6 +4,7 @@ import javax.persistence.*;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 @Table(name = "menu")
 @Entity
