+++ src/main/java/kitchenpos/menugroup/domain/MenuGroupRepository.java
@@ -1,8 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.menugroup.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 public interface MenuGroupRepository {
     MenuGroup save(MenuGroup menuGroup);
