+++ src/test/java/kitchenpos/Fixtures.java
@@ -7,6 +7,7 @@ import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.Random;
 import java.util.UUID;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 public class Fixtures {
     public static final UUID INVALID_ID = new UUID(0L, 0L);
